January Update #2
Corrected an error in the isochronal apd calculation. Addressed the issue of resetting GUI features when the "Load" button is clicked a second time or the analysis method is changed. Updated the fluorescence ratio calculations to operate on the raw data signal. Made APD Batch checkboxes default to "CHECKED" once a signal has been selected for the axes. Reduced redundant instances of code substantially through use of string formatting and the "getattr" command (e.g., lines 956 & 957).